Fix up for windows pointer sizes.
As title.

+++ numba/_helperlib.c
@@ -1064,7 +1064,8 @@ numba_unpickle(const char *data, int n)
  */
 
 NUMBA_EXPORT_FUNC(void *)
-numba_extract_unicode(PyObject *obj, Py_ssize_t *length, int *kind, long *hash) {
+numba_extract_unicode(PyObject *obj, Py_ssize_t *length, int *kind,
+                      Py_ssize_t *hash) {
 #if (PY_MAJOR_VERSION >= 3) && (PY_MINOR_VERSION >= 3)
     if (!PyUnicode_READY(obj)) {
         *length = PyUnicode_GET_LENGTH(obj);
